// integrations/netsuite-tba/schema.ts
export const netsuiteInvoiceUpdateInputSchema = z.object({
    id: z.string().min(1),
    customerId: z.string().min(1).optional(),
    currency: z.string().min(1).optional(),
    description: z.string().optional(),
    status: z.string().min(1).optional(),
    lines: z
        .array(
            z.object({
                itemId: z.string().min(1),
                quantity: z.number().min(0),
                amount: z.number().min(0),
                locationId: z.string().optional(),
                vatCode: z.string().optional(),
                description: z.string().optional()
            })
        )

        .optional()
});